Prototype nearly finished
Has Traps and moving pieces implemented. Need to get a working death plane and fix respawing. Also would like to be able to add the track to main menu level selectin so We can have enemies as well. But I am worried they will tweak out with the updated track