Reflector Shield Fix/Powers Added to All Ships
I fixed the reflector shield by getting the enabling/disabling to work correctly. I then added the powers to all other ships so that when any ship runs over a powerup pickup, it will randomly receive a powerup that it can use. Ships can only use one power up at a time. These were all made into new prefabs that are in the folder with SethBennett in the title under the prefabs folder.