Fix Powerup Bug
I added a little bit of code that turns the sphere collider on and off with the powerup to the EMP. To do this, I had to disable the sphere collider on each prefab. For a secondary safety precaution, I also added a distance detector to check for the pickup. These combined should make it to where the powerups are only picked up when the ship flies through it, nothing else. I also changed the if statement in the Trap script. It was using a "tag" call, I changed it to a layer call. This seems to have fixed the collision detection problem, similar to the powerup collision detection problem.